requirements.txt
pandas
python-dotenv
fsspec
openpyxl